Indoor staging

+++ resources/views/my-account/my-account-scripts.blade.php
@@ -745,8 +745,7 @@
 
                         var product_id = item.prd_id;
                         var price = parseFloat(item.price);
-                        var quantity = parseFloat(item.quantity);
-
+                        var quantity = parseInt(item.quantity);
                         var subtotal = parseFloat(price * quantity);
                         var cart_total = cart_total + subtotal;
                         var total_cart_quantity = total_cart_quantity + quantity;
@@ -869,8 +868,7 @@
 
                         var product_id = item.prd_id;
                         var price = parseFloat(item.price);
-                        var quantity = parseFloat(item.quantity);
-
+                        var quantity = parseInt(item.quantity);
                         var subtotal = parseFloat(price * quantity);
                         var cart_total = cart_total + subtotal;
                         var total_cart_quantity = total_cart_quantity + quantity;
@@ -899,7 +897,6 @@
                         var product_id = item.prd_id;
                         var price = parseFloat(item.price);
                         var quantity = parseInt(item.quantity);
-
                         var subtotal = parseFloat(price * quantity);
                         var cart_total = cart_total + subtotal;
                         var total_cart_quantity = total_cart_quantity + quantity;
@@ -1051,8 +1048,7 @@
 
                         var product_id = item.prd_id;
                         var price = parseFloat(item.price);
-                        var quantity = parseFloat(item.quantity);
-
+                        var quantity = parseInt(item.quantity);
                         var subtotal = parseFloat(price * quantity);
                         var cart_total = cart_total + subtotal;
                         var total_cart_quantity = total_cart_quantity + quantity;
